requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
Django==3.0.8
django-cors-headers==3.4.0
mysqlclient==2.0.1
pudb==2019.2
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pytz==2020.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
urwid==2.1.0
haversine==2.2.0
requests==2.24.0
django-extensions==3.0.3
ipython==7.16.1